Split object persistence parser
